+++ app/Http/Controllers/ArtistaController.php
@@ -2,7 +2,9 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\AlbumArtista;
 use App\Models\Artista;
+use App\Models\ArtistaCancion;
 use Illuminate\Http\Request;
 
 class ArtistaController extends Controller
@@ -81,6 +83,16 @@ class ArtistaController extends Controller
      */
     public function destroy(Artista $artista)
     {
+        //ArtistaCancion::where('artista_id', $artista->id)->delete();
+        //AlbumArtista::where('artista_id', $artista->id)->delete();   //esto es si queremos que se borre y listo, pero el ejercicio nos dice que se impida borrar
+        $r1 = ArtistaCancion::where('artista_id', $artista->id)->count();
+        $r2 = AlbumArtista::where('artista_id', $artista->id)->count();
+
+        if ($r1 > 0 || $r2 > 0) {
+            session()->flash('error', 'No se puede borrar una artista contenido en un álbum o en una canción');
+            return redirect()->route('canciones.index');   //sin esto no funcionaría
+        }
+
         $artista->delete();
         session()->flash('success', 'Artista eliminado correctamente.');
         return redirect()->route('artistas.index');
